+++ changelog.md
@@ -76,6 +76,7 @@
     - Fixed do not escape disabled special characters
     - Fixed command character not being correctly replaced with custom character in some places
     - Fixed saving nested loop variables as user variables
+    - Fixed bug when splitting quoted strings
   - Fixed [%dice(xdy+n)](docs/functions.md#math) not correctly parsings argument format
   - Fixed [#variable](docs/commands.md) should not evaluate the value to allow use of raw strings as needed, eg #va test {3+5} should store 3+5 not 8, if you want the evaluated value use #var test \${3+5} or #va test \{${3+5}}
   - Fixed [#variable](docs/commands.md) throws an error now when invalid variable names are used